This PR was merged into the 1.5-dev branch.

Discussion
----------



Commits
-------

64b7a4689397cfa390a5863c9cc008e27e19696f Introduce Psalm

+++ src/Translation/Finder/OrderingTranslationFilesFinder.php
@@ -31,7 +31,7 @@ final class OrderingTranslationFilesFinder implements TranslationFilesFinderInte
             $firstFile = str_replace($path, '', $firstFile);
             $secondFile = str_replace($path, '', $secondFile);
 
-            return strpos($secondFile, 'translations') - strpos($firstFile, 'translations');
+            return (int) strpos($secondFile, 'translations') - (int) strpos($firstFile, 'translations');
         });
 
         return $files;
